Add a base class for all continuous read devices
